Updated ProbabilityUpdates
ProbabilityUpdates2.py now checks to see if POAsum has a value before
proceeding.  If it does not than it most likely has not been assigned
yet.  If not assign, the script checks to see if values exist in
POCaSsign than calculate POAcum.  If they do exist than this initial
calculation is bypassed and the conventional update proceeds.