Probability Updates #2
When the Probability Updates is run the script now confirms if the
Probabilities had been previously entered.  If not it calculates the POA
based on the assigned values and normalizes.  If the POA has already
been assigned and normalized, the script re-calculates the value and
compares to the existing value in case a change was made to the assigned
POA.  Then the script looks for any updates through Assignment_Debrief.