Automate planning numbers - update Assignment designation
The QRTSegmentAssignments now given the option of auto-generating the
Planning Number.  Updates were also made to the Tool Validator to ensure
the parameters fields are cleared out appropriately when the tool is
initiated.

Also added a "Message" to the user when the Update Layout and New
Incident tools are run reminding them that if they are using a Projected
coordinate system they should turn on the ASRC USNG grid.