Removed special characters from file name attempts
When the cell_Coverage tool was executed it would look at the
"DESCRIPTION" field in the cell tower layer for the name to assign the
feature and raster results of the tool.  If the DESCRIPTION had special
characters it would throw an error when assigned as a file name.  Added
module import for Regular Expressions to replace any special characters
in the DESCRIPTION when assigning the feature name.