Grid north correction and updated map templates.
Added functionality in UpdateLayout.py to calculate grid north/true
north correction based on IPP or ICP and rotate map in layout to True
North.

When the Create Assignment tool is executed the user now has the option
to update the map layout prior to generating task maps.

Updated all of the available Templates.
Updated