Updated scripts and add XY Point Plot tool
Updated all of the scripts to only load arcpy, sys and math modules only
if not already present.  Also included a new script for plotting the
location of a point feature based on the coordinate format and any
coordinate system used when reported - could be different than data
frame coordinate system.