Modifications to support automatic rewriting of bearing conversion text on all map templates.  The static text that was there was only correct for negative (west) declination.
Since UpdateLayout is already computing the declination and adjusting
the cardinal direction of declination based on sign, it is also possible to
tweak the conversion text based on the sign.

This required changing *all* map templates so that the text element
containing the bearing conversion text has a name, which I set to be
"bearingConv".  Then UpdateLayout can pull  out that string and modify it.

My original attempt was just to have the string contain "%s" instead of
"add" or "subtract", and then just use string substition to populate with
the correct words.  But this was an incomplete solution --- once
UpdateLayout is run, the string no longer has the %s in it.  So one has
to put them back each time.