Check for ICS204 and update Segment_Points for task maps
Added check in IGT4SAR_CreateICS204.py to look for the designated 204
form based on user defined value in IGT4SAR_CreateAssignment_gpx.py.

Updated how Segment_Points are added to the task assignment map.  There
was an error that was resulting in all of the points being included not
just the ones that were within 20 meters of the assigned feature.