Major overhaul related to ICS204 Forms
Added a text file in C:|MapSAR_Ex\Forms\TAF_ICS204 called
"ICS204Forms_Available.txt".  This text file has a list of available pdf
forms (column 2) and descriptive index key (column 1).  This is now the
only file that would need modified (aside from mapping fields in
IGT4SAR_CreateICS204.py) to include new forms.

Modified Create New Incident script to ask the use which ICS204 form
they wish to use during the creation of a new incident.  The indicated
Key Index value is stored in the Incident Information Table under the
field name of ICS204.  The user can modify this entry if they wish to
use multiple forms.

When the CreateAssignment_gpx.py script is run it first identifies the
desired form to use by looking at the key value in the incident
information.  Next the script checks to see if the script is available
in the Assignments folder of your incident.  If it is not, it attempts
to get a copy from the Forms\TAF_ICS204 folder.